+++ Examples/Resources/SqlAG/1-CreateAvailabilityGroup.ps1
@@ -32,8 +32,8 @@ Configuration Example
 {
     param(
         [Parameter(Mandatory = $true)]
-        [PSCredential]
-        $SysAdminAccount
+        [System.Management.Automation.PSCredential]
+        $SqlAdministratorCredential
     )
 
     Import-DscResource -ModuleName SqlServerDsc
@@ -47,7 +47,7 @@ Configuration Example
             LoginType            = 'WindowsUser'
             ServerName           = $Node.NodeName
             InstanceName         = $Node.InstanceName
-            PsDscRunAsCredential = $SysAdminAccount
+            PsDscRunAsCredential = $SqlAdministratorCredential
         }
 
         # Add the required permissions to the cluster service login
@@ -59,7 +59,7 @@ Configuration Example
             InstanceName         = $Node.InstanceName
             Principal            = 'NT SERVICE\ClusSvc'
             Permission           = 'AlterAnyAvailabilityGroup', 'ViewServerState'
-            PsDscRunAsCredential = $SysAdminAccount
+            PsDscRunAsCredential = $SqlAdministratorCredential
         }
 
         # Create a DatabaseMirroring endpoint
@@ -70,7 +70,7 @@ Configuration Example
             Port                 = 5022
             ServerName           = $Node.NodeName
             InstanceName         = $Node.InstanceName
-            PsDscRunAsCredential = $SysAdminAccount
+            PsDscRunAsCredential = $SqlAdministratorCredential
         }
 
         if ( $Node.Role -eq 'PrimaryReplica' )
@@ -83,7 +83,7 @@ Configuration Example
                 InstanceName         = $Node.InstanceName
                 ServerName           = $Node.NodeName
                 DependsOn            = '[SqlServerEndpoint]HADREndpoint', '[SqlServerPermission]AddNTServiceClusSvcPermissions'
-                PsDscRunAsCredential = $SysAdminAccount
+                PsDscRunAsCredential = $SqlAdministratorCredential
             }
         }
     }