+++ Examples/Resources/SqlDatabaseRole/1-AddDatabaseRole.ps1
@@ -9,8 +9,7 @@ Configuration Example
     (
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
-        [System.Management.Automation.Credential()]
-        $SysAdminAccount
+        $SqlAdministratorCredential
     )
 
     Import-DscResource -ModuleName SqlServerDsc
@@ -25,7 +24,7 @@ Configuration Example
             Name                 = 'CONTOSO\SQLAdmin'
             Role                 = 'MyRole', 'MySecondRole'
             Database             = 'AdventureWorks'
-            PsDscRunAsCredential = $SysAdminAccount
+            PsDscRunAsCredential = $SqlAdministratorCredential
         }
     }
 }