+++ Examples/Resources/SqlScript/1-RunScriptUsingSQLAuthentication.ps1
@@ -7,7 +7,7 @@ Configuration Example
 {
     param(
         [Parameter(Mandatory = $true)]
-        [PSCredential]
+        [System.Management.Automation.PSCredential]
         $SqlCredential
     )