+++ Examples/Resources/SqlServerEndpointState/2-MakeSureEndpointIsStopped.ps1
@@ -8,8 +8,7 @@ Configuration Example
     (
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
-        [System.Management.Automation.Credential()]
-        $SysAdminAccount
+        $SqlAdministratorCredential
     )
 
     Import-DscResource -ModuleName SqlServerDsc
@@ -23,7 +22,7 @@ Configuration Example
             Name                 = 'DefaultMirrorEndpoint'
             State                = 'Stopped'
 
-            PsDscRunAsCredential = $SysAdminAccount
+            PsDscRunAsCredential = $SqlAdministratorCredential
 
         }
     }