+++ Examples/Resources/SqlServerRole/1-AddServerRole.ps1
@@ -8,7 +8,7 @@ Configuration Example
 {
     param(
         [Parameter(Mandatory = $true)]
-        [PSCredential]
+        [System.Management.Automation.PSCredential]
         $SysAdminAccount
     )