+++ Examples/Resources/SqlSetup/4-InstallNamedInstanceInFailoverClusterFirstNode.ps1
@@ -37,25 +37,21 @@ Configuration Example
         [Parameter(Mandatory = $true)]
         [ValidateNotNullorEmpty()]
         [System.Management.Automation.PSCredential]
-        [System.Management.Automation.Credential()]
         $SqlInstallCredential,
 
         [Parameter()]
         [ValidateNotNullorEmpty()]
         [System.Management.Automation.PSCredential]
-        [System.Management.Automation.Credential()]
         $SqlAdministratorCredential = $SqlInstallCredential,
 
         [Parameter(Mandatory = $true)]
         [ValidateNotNullorEmpty()]
         [System.Management.Automation.PSCredential]
-        [System.Management.Automation.Credential()]
         $SqlServiceCredential,
 
         [Parameter()]
         [ValidateNotNullorEmpty()]
         [System.Management.Automation.PSCredential]
-        [System.Management.Automation.Credential()]
         $SqlAgentServiceCredential = $SqlServiceCredential
     )