+++ CHANGELOG.md
@@ -35,6 +35,9 @@
   - Resources that did not have a description in the README.md now has one.
   - Resources that missed links to the examples in the README.md now has those
     links.
+  - Style changes in all examples, removing type [System.Management.Automation.Credential()]
+    from credential parameters ([issue #1003](https://github.com/PowerShell/SqlServerDsc/issues/1003)),
+    and renamed the credential parameter so it is not using abbreviation.
 - Changes to SqlAlias
   - Fixed issue where exception was thrown if reg keys did not exist
     ([issue #949](https://github.com/PowerShell/SqlServerDsc/issues/949)).